+++ BusinessObjects/Dtos/Bids/CreateBidRequest.cs
@@ -5,6 +5,6 @@ namespace BusinessObjects.Dtos.Bids;
 public class CreateBidRequest
 {
     public Guid MemberId { get; set; }
-    [Range(0,int.MaxValue,ErrorMessage = "Amount must be greater than 0")]
-    public int Amount { get; set; }
+    [Range(0,(double)decimal.MaxValue,ErrorMessage = "Amount must be greater than 0")]
+    public decimal Amount { get; set; }
 }
