+++ BusinessObjects/Dtos/FashionItems/UpdateFashionItemRequest.cs
@@ -10,7 +10,7 @@ namespace BusinessObjects.Dtos.FashionItems
 {
     public class UpdateFashionItemRequest
     {
-        public int? SellingPrice { get; set; }
+        public decimal? SellingPrice { get; set; }
         public string? Name { get; set; }
         public string? Note { get; set; }
