+++ BusinessObjects/Dtos/Refunds/RefundResponse.cs
@@ -25,7 +25,7 @@ namespace BusinessObjects.Dtos.Refunds
         public string CustomerEmail { get; set; }
         public string? ResponseFromShop { get; set; }
         public int? RefundPercentage { get; set; }
-        public int? RefundAmount { get; set; }
+        public decimal? RefundAmount { get; set; }
 
         public string[] Images { get; set; }
         public RefundStatus RefundStatus { get; set; }
